Moves project related code from occlient.go to a new file
